+++ packages/list/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.21.76](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.75...@ant-design/pro-list@1.21.76) (2022-07-22)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [1.21.75](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.74...@ant-design/pro-list@1.21.75) (2022-07-21)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [1.21.74](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.73...@ant-design/pro-list@1.21.74) (2022-07-21)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
 ## [1.21.73](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.72...@ant-design/pro-list@1.21.73) (2022-07-14)
 
 **Note:** Version bump only for package @ant-design/pro-list